Isolate database auth from middleware
