+++ src/auth.ts
@@ -1,12 +1,15 @@
 import NextAuth from "next-auth";
 import { DrizzleAdapter } from "@auth/drizzle-adapter";
+import Credentials from "next-auth/providers/credentials";
+import bcrypt from "bcryptjs";
 import { db } from "./server/db";
 
-import { getUserById } from "./server/queries/users";
+import { getUserByEmail, getUserById } from "./server/queries/users";
 import authConfig from "~/auth.config";
 import { users } from "./server/db/schema";
 import { eq } from "drizzle-orm";
 import { getAccountByUserId } from "./server/queries/account";
+import { LoginSchema } from "./schemas";
 
 export const {
   handlers: { GET, POST },
@@ -14,6 +17,29 @@ export const {
   signIn,
   signOut,
 } = NextAuth({
+  ...authConfig,
+  providers: [
+    ...authConfig.providers,
+    Credentials({
+      async authorize(credentials) {
+        const validatedFields = LoginSchema.safeParse(credentials);
+
+        if (validatedFields.success) {
+          const { email, password } = validatedFields.data;
+
+          const user = await getUserByEmail(email);
+
+          if (!user || !user.password) return null;
+
+          const passwordMatch = await bcrypt.compare(password, user.password);
+
+          if (passwordMatch) return user;
+        }
+
+        return null;
+      },
+    }),
+  ],
   pages: {
     signIn: "/auth/signin",
     error: "/auth/error",
@@ -74,5 +100,4 @@ export const {
   },
   adapter: DrizzleAdapter(db),
   session: { strategy: "jwt" },
-  ...authConfig,
 });
